Remove fallback to PSM 1 on connection error
Previously, if the one-time check fails, the fuzzer changes target port
to 1. This may not be clear to user, and it's better to allow user to
decide what target to fuzz! imo.
Besides, no check was/is done to ensure that the new port 1 communication is successful,
which creates confusing patterns, like:
"cannot communicate with port 1! Defaulting to port 1".